requirements.txt
aiogram==3.15
python-dotenv
